+++ src/constants.js
@@ -5,7 +5,7 @@
  */
 
 /** Pulse Card version string. */
-export const VERSION = '1.2.1';
+export const VERSION = '1.3.0';
 
 /** Console log prefix. */
 export const LOG_PREFIX = 'Pulse Card:';